WIP: dts: support having DEFAULT_DEVICE_TREE in U-Boot dir with OF_UPSTREAM
For the multi-DTB FIT usecase it may still be desirable to have a dummy
default DTB, which could be used to define additional match data and/or
contain info like the memory layout which is filled in by a previous
stage bootloader.

These DTs are highly specific to U-Boot and so would be kept under the
arch/$(ARCH)/dts/ directory. Add support for building these when
CONFIG_OF_UPSTREAM is set.

FIXME: this is still quite finnicky ...

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>