clk/stub: add sc7280-rpmhpd PD stub
The RPMh PD is a vote-driven power-controller provided by the RPMh
co-processor. Since we aren't implementing power management support in
U-Boot (e.g. for entering low power states) we can just stub this and
rely on the votes provided by earlier boot stages.

Add a specific of_match to the PD stub to match against this for SC7280.
This is required for i2c and spi.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>